refactor: separate workload types in `cli/deploy`
